requirements.txt
gunicorn
Flask
Flask-Bootstrap
Flask-Mail
Flask-SQLAlchemy
Flask-MySQL
Flask-BasicAuth
Flask-Security
utm
python-memcached
requests
bson
sqlalchemy-migrate
lxml
cssselect
python-dateutil
